+++ src/util/session/storageBuilder.ts
@@ -1,5 +1,4 @@
 import { createCookieSessionStorage, createSessionStorage, SessionData, SessionIdStorageStrategy } from "@remix-run/node";
-import { getEnv } from "../getEnv";
 
 /**
  * Represents a method used to create entries in the database.
@@ -34,10 +33,8 @@ interface generateCookieProps {
 
     /**
      * Secret for the stored cookies.
-     *
-     * @defaultValue process.env.COOKIE_AUTH_SECRET
      */
-    "secret"?: string
+    "secret": string
 }
 
 /**
@@ -73,7 +70,7 @@ export interface databaseBuilderProps extends generateCookieProps {
 /**
  * Generates a cookie object.
  */
-function generateCookie({ name = "_session", secret = getEnv("COOKIE_AUTH_SECRET") }: generateCookieProps): SessionIdStorageStrategy["cookie"] {
+function generateCookie({ name = "_session", secret }: generateCookieProps): SessionIdStorageStrategy["cookie"] {
     if (!secret) {
         throw new Error("'secret' must not be undefined.");
     }